# class_2_VEprojectORIGINAL/module_1_VEproject/app/static/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Mechanic Shop"
  description: "Module 2 for API"
host: "localhost:5000"
basePath: "/"
schemes:
  - "http"
securityDefinitions:
  BearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
paths:
...
  /service_tickets:
    get:
      tags:
        - "Service Tickets"
      summary: "Retrieve all service tickets"
      responses:
        200:
          description: "List of service tickets."
        500:
          description: "Server error."
    post:
      tags:
        - "Service Tickets"
      summary: "Create a new service ticket"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/ServiceTicket"
      responses:
        201:
          description: "Service ticket created successfully."
        400:
          description: "Validation error."
  /service_tickets/{id}:
    put:
      tags:
        - "Service Tickets"
      summary: "Update a service ticket's details"
      parameters:
        - name: "id"
          in: "path"
          type: "integer"
          required: true
          description: "Service ticket ID."
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/ServiceTicket"
      responses:
        200:
          description: "Service ticket updated successfully."
        404:
          description: "Service ticket not found."
        400:
          description: "Validation error."
    delete:
      tags:
        - "Service Tickets"
      summary: "Delete a service ticket"
      parameters:
        - name: "id"
          in: "path"
          type: "integer"
          required: true
          description: "Service ticket ID."
      responses:
        200:
          description: "Service ticket deleted successfully."
        404:
          description: "Service ticket not found."
definitions:
...
  ServiceTicket:
    type: object
    required:
      - service_description
      - cost
      - vin_number
      - work_complete
      - car_submission_date
      - customer_id
      - mechanic_id
    properties:
      id:
        type: integer
        readOnly: true
      service_description:
        type: string
      cost:
        type: number
      vin_number:
        type: string
      work_complete:
        type: boolean
      car_submission_date:
        type: string
        format: date
      work_start_date:
        type: string
        format: date
      work_finish_date:
        type: string
        format: date
      customer_id:
        type: integer
      mechanic_id:
        type: integer